+++ test/vector/vml/sv_mul.hpp
@@ -17,7 +17,7 @@ template <typename T> bool test_send_svmul(const size_t size, double tol) {
   monolish::vector<T> x(size, 0.1, 1.0);
   monolish::vector<T> ans(size, 321.0);
 
-  monolish::vector<T> ans_tmp = ans.copy();
+  monolish::vector<T> ans_tmp(ans);
   ans_svmul(x, value, ans_tmp);
 
   monolish::util::send(x, ans);
@@ -33,7 +33,7 @@ template <typename T> bool test_svmul(const size_t size, double tol) {
   monolish::vector<T> x(size, 0.1, 1.0);
   monolish::vector<T> ans(size, 321.0);
 
-  monolish::vector<T> ans_tmp = ans.copy();
+  monolish::vector<T> ans_tmp(ans);
   ans_svmul(x, value, ans_tmp);
 
   monolish::vml::mul(x, value, ans);