+++ test/vector/vml/v_reciprocal.hpp
@@ -11,8 +11,7 @@ bool test_send_vreciprocal(const size_t size, double tol) {
 
   monolish::vector<T> ans(size, 0.1, 10.0);
 
-  monolish::vector<T> ans_tmp;
-  ans_tmp = ans.copy();
+  monolish::vector<T> ans_tmp(ans);
   ans_vreciprocal(ans_tmp);
 
   monolish::util::send(ans);
@@ -26,8 +25,7 @@ template <typename T> bool test_vreciprocal(const size_t size, double tol) {
 
   monolish::vector<T> ans(size, 0.1, 10.0);
 
-  monolish::vector<T> ans_tmp;
-  ans_tmp = ans.copy();
+  monolish::vector<T> ans_tmp(ans);
   ans_vreciprocal(ans_tmp);
 
   monolish::vml::reciprocal(ans, ans);